+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Refactor reducer to accept a context argument
+
 ## 2.1.1 (2019-04-29)
 
 * Fix an issue where the placementEngine was initially emitting _all_ promos to subscribers, before filtering them